moved styling and docker postgres port
I was having a problem where I could not see the docker database in PGAdmin, so I decided to move it to another port. The alternative would be making sure postgres is stopped before every docker compose, so I might switch to that if I feel that having two identical databases is not working for me.